refactor: move account balance code to new file
